split SynthesisEngine mora utility to mora module
