+++ voicevox_engine/mora/mora_to_text.py
@@ -1,0 +1,10 @@
+from ..mora_list import openjtalk_mora2text
+
+def mora_to_text(mora: str) -> str:
+    if mora[-1:] in ["A", "I", "U", "E", "O"]:
+        # 無声化母音を小文字に
+        mora = mora[:-1] + mora[-1].lower()
+    if mora in openjtalk_mora2text:
+        return openjtalk_mora2text[mora]
+    else:
+        return mora
