+++ voicevox_engine/mora/split_mora.py
@@ -1,0 +1,42 @@
+from typing import List, Optional
+
+from ..acoustic_feature_extractor import OjtPhoneme
+
+unvoiced_mora_phoneme_list = ["A", "I", "U", "E", "O", "cl", "pau"]
+mora_phoneme_list = ["a", "i", "u", "e", "o", "N"] + unvoiced_mora_phoneme_list
+
+
+def split_mora(phoneme_list: List[OjtPhoneme]):
+    """
+    OjtPhonemeのリストから、
+    母音の位置(vowel_indexes)
+    母音の音素列(vowel_phoneme_list)
+    子音の音素列(consonant_phoneme_list)
+    を生成し、返す
+    Parameters
+    ----------
+    phoneme_list : List[OjtPhoneme]
+        phonemeクラスのリスト
+    Returns
+    -------
+    consonant_phoneme_list : List[OjtPhoneme]
+        子音の音素列
+    vowel_phoneme_list : List[OjtPhoneme]
+        母音の音素列
+    vowel_indexes : : List[int]
+        母音の位置
+    """
+    vowel_indexes = [
+        i for i, p in enumerate(phoneme_list) if p.phoneme in mora_phoneme_list
+    ]
+    vowel_phoneme_list = [phoneme_list[i] for i in vowel_indexes]
+    # postとprevのvowel_indexの差として考えられる値は1か2
+    # 理由としてはphoneme_listは、consonant、vowelの組み合わせか、vowel一つの連続であるから
+    # 1の場合はconsonant(子音)が存在しない=母音のみ(a/i/u/e/o/N/cl/pau)で構成されるモーラ(音)である
+    # 2の場合はconsonantが存在するモーラである
+    # なので、2の場合(else)でphonemeを取り出している
+    consonant_phoneme_list: List[Optional[OjtPhoneme]] = [None] + [
+        None if post - prev == 1 else phoneme_list[post - 1]
+        for prev, post in zip(vowel_indexes[:-1], vowel_indexes[1:])
+    ]
+    return consonant_phoneme_list, vowel_phoneme_list, vowel_indexes
