+++ voicevox_engine/synthesis_engine/SynthesisEngine.py
@@ -7,133 +7,10 @@ from scipy.signal import resample
 from ..acoustic_feature_extractor import OjtPhoneme, SamplingData
 from ..full_context_label import extract_full_context_label
 from ..model import AccentPhrase, AudioQuery, Mora
-from ..mora_list import openjtalk_mora2text
 
-unvoiced_mora_phoneme_list = ["A", "I", "U", "E", "O", "cl", "pau"]
-mora_phoneme_list = ["a", "i", "u", "e", "o", "N"] + unvoiced_mora_phoneme_list
-
-
-# TODO: move mora utility to mora module
-def to_flatten_moras(accent_phrases: List[AccentPhrase]) -> List[Mora]:
-    """
-    accent_phrasesに含まれるMora(とpause_moraがあればそれも)を
-    すべて一つのリストに結合する
-    Parameters
-    ----------
-    accent_phrases : List[AccentPhrase]
-        AccentPhraseのリスト
-    Returns
-    -------
-    moras : List[Mora]
-        結合されたMoraのリストを返す
-    """
-    return list(
-        chain.from_iterable(
-            accent_phrase.moras
-            + (
-                [accent_phrase.pause_mora]
-                if accent_phrase.pause_mora is not None
-                else []
-            )
-            for accent_phrase in accent_phrases
-        )
-    )
-
-
-def to_phoneme_data_list(phoneme_str_list: List[str]):
-    """
-    phoneme文字列のリストを、OjtPhonemeクラスのリストに変換する
-    Parameters
-    ----------
-    phoneme_str_list : List[str]
-        phoneme文字列のリスト
-    Returns
-    -------
-    phoneme_list : List[OjtPhoneme]
-        変換されたOjtPhonemeクラスのリスト
-    """
-    phoneme_data_list = [
-        OjtPhoneme(phoneme=p, start=i, end=i + 1)
-        for i, p in enumerate(phoneme_str_list)
-    ]
-    phoneme_data_list = OjtPhoneme.convert(phoneme_data_list)
-    return phoneme_data_list
-
-
-def split_mora(phoneme_list: List[OjtPhoneme]):
-    """
-    OjtPhonemeのリストから、
-    母音の位置(vowel_indexes)
-    母音の音素列(vowel_phoneme_list)
-    子音の音素列(consonant_phoneme_list)
-    を生成し、返す
-    Parameters
-    ----------
-    phoneme_list : List[OjtPhoneme]
-        phonemeクラスのリスト
-    Returns
-    -------
-    consonant_phoneme_list : List[OjtPhoneme]
-        子音の音素列
-    vowel_phoneme_list : List[OjtPhoneme]
-        母音の音素列
-    vowel_indexes : : List[int]
-        母音の位置
-    """
-    vowel_indexes = [
-        i for i, p in enumerate(phoneme_list) if p.phoneme in mora_phoneme_list
-    ]
-    vowel_phoneme_list = [phoneme_list[i] for i in vowel_indexes]
-    # postとprevのvowel_indexの差として考えられる値は1か2
-    # 理由としてはphoneme_listは、consonant、vowelの組み合わせか、vowel一つの連続であるから
-    # 1の場合はconsonant(子音)が存在しない=母音のみ(a/i/u/e/o/N/cl/pau)で構成されるモーラ(音)である
-    # 2の場合はconsonantが存在するモーラである
-    # なので、2の場合(else)でphonemeを取り出している
-    consonant_phoneme_list: List[Optional[OjtPhoneme]] = [None] + [
-        None if post - prev == 1 else phoneme_list[post - 1]
-        for prev, post in zip(vowel_indexes[:-1], vowel_indexes[1:])
-    ]
-    return consonant_phoneme_list, vowel_phoneme_list, vowel_indexes
-
-
-def pre_process(
-    accent_phrases: List[AccentPhrase],
-) -> Tuple[List[Mora], List[OjtPhoneme]]:
-    """
-    AccentPhraseモデルのリストを整形し、処理に必要なデータの原型を作り出す
-    Parameters
-    ----------
-    accent_phrases : List[AccentPhrase]
-        AccentPhraseモデルのリスト
-    Returns
-    -------
-    flatten_moras : List[Mora]
-        AccentPhraseモデルのリスト内に含まれるすべてのMoraをリスト化したものを返す
-    phoneme_data_list : List[OjtPhoneme]
-        flatten_morasから取り出したすべてのPhonemeをOjtPhonemeに変換したものを返す
-    """
-    flatten_moras = to_flatten_moras(accent_phrases)
-
-    phoneme_each_mora = [
-        ([mora.consonant] if mora.consonant is not None else []) + [mora.vowel]
-        for mora in flatten_moras
-    ]
-    phoneme_str_list = list(chain.from_iterable(phoneme_each_mora))
-    phoneme_str_list = ["pau"] + phoneme_str_list + ["pau"]
-
-    phoneme_data_list = to_phoneme_data_list(phoneme_str_list)
-
-    return flatten_moras, phoneme_data_list
-
-
-def mora_to_text(mora: str) -> str:
-    if mora[-1:] in ["A", "I", "U", "E", "O"]:
-        # 無声化母音を小文字に
-        mora = mora[:-1] + mora[-1].lower()
-    if mora in openjtalk_mora2text:
-        return openjtalk_mora2text[mora]
-    else:
-        return mora
+from ..mora.mora_to_text import mora_to_text
+from ..mora.pre_process import pre_process
+from ..mora.split_mora import split_mora, unvoiced_mora_phoneme_list
 
 
 class SynthesisEngine:
